refactor(api): move Redis cache file and test
