+++ api/src/shared/infrastructure/caches/learning-content-cache.js
@@ -1,8 +1,8 @@
-import { RedisCache } from '../../../../lib/infrastructure/caches/RedisCache.js';
 import { config } from '../../config.js';
 import { DistributedCache } from './DistributedCache.js';
 import { InMemoryCache } from './InMemoryCache.js';
 import { LayeredCache } from './LayeredCache.js';
+import { RedisCache } from './RedisCache.js';
 
 const LEARNING_CONTENT_CHANNEL = 'Learning content';
 const LEARNING_CONTENT_CACHE_KEY = 'LearningContent';
